moved all client code to /client and refactored out components to their respective .vue files
