+++ client/components.js
@@ -1,0 +1,3 @@
+console.log('loading components');
+
+console.log('delete this file after merge with vue-refactor');
