+++ public/timers.html
@@ -9,7 +9,7 @@
 
     <script src="/javascripts/timings.bundle.js"></script>
 </head>
-<body onload="initAppz()">
+<body onload="window.bootstrapTimingsApp()">
 <div class="btn-danger" style="text-align:center; font-size: 150%;">Page only support viewing in Google Chrome at the
     moment
 </div>
